chore:fetch함수 파일 경로 수정
