+++ src/apis/auth.ts
@@ -43,3 +43,14 @@ export const fetchUserInfoDetail = async (accessToken: string): Promise<UserInfo
 const authApi = { requestTokens, requestUserInfo, fetchUserInfoDetail };
 
 export default authApi;
+
+export const fetchUserInfoDetail = async (
+  accessToken: string
+): Promise<UserInfoResponse> => {
+  const response = await request.get('/users/info/detail', {
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+    },
+  });
+  return response.data;
+};
